Update hearing details page to accept a DTO and book through the journey until the summary page
