+++ UI.PageModels/Pages/Admin/Booking/HearingDetailsPage.cs
@@ -1,4 +1,5 @@
 ﻿using OpenQA.Selenium;
+using UI.PageModels.Dtos;
 
 namespace UI.PageModels.Pages.Admin.Booking;
 
@@ -17,6 +18,13 @@ public class HearingDetailsPage : VhAdminWebPage
                 "This is not the HearingDetails page, the current url is: " + Driver.Url);
     }
 
+    /// <summary>
+    /// Enter the case details, case type and hearing type
+    /// </summary>
+    /// <param name="caseNumber"></param>
+    /// <param name="caseName"></param>
+    /// <param name="caseType"></param>
+    /// <param name="hearingType"></param>
     public void EnterHearingDetails(string caseNumber, string caseName, string caseType, string hearingType)
     {
         WaitForApiSpinnerToDisappear();
@@ -31,5 +39,33 @@ public class HearingDetailsPage : VhAdminWebPage
     {
         ClickElement(_nextButton);
         return new HearingSchedulePage(Driver, DefaultWaitTime);
+    }
+
+    /// <summary>
+    /// Enter the details for a hearing and go to the summary page
+    /// </summary>
+    /// <param name="bookingDto"></param>
+    /// <returns></returns>
+    public SummaryPage EnterHearingDetails(BookingDto bookingDto)
+    {
+        EnterHearingDetails(bookingDto.CaseNumber, bookingDto.CaseName, bookingDto.CaseType,
+            bookingDto.HearingType);
+        var hearingSchedulePage = GoToNextPage();
+
+        hearingSchedulePage.EnterSingleDayHearingSchedule(bookingDto.ScheduledDateTime, bookingDto.DurationHour,
+            bookingDto.DurationMinute, bookingDto.VenueName, bookingDto.RoomName);
+
+        var assignJudgePage = hearingSchedulePage.GoToNextPage();
+        assignJudgePage.EnterJudgeDetails(bookingDto.Judge.Username, bookingDto.Judge.DisplayName, bookingDto.Judge.Phone);
+
+        var addParticipantPage = assignJudgePage.GoToParticipantsPage();
+        addParticipantPage.AddExistingParticipants(bookingDto.Participants);
+        var videoAccessPointsPage = addParticipantPage.GoToVideoAccessPointsPage();
+        videoAccessPointsPage.AddVideoAccessPoints(bookingDto.VideoAccessPoints);
+        var otherInformationPage = videoAccessPointsPage.GoToOtherInformationPage();
+        otherInformationPage.TurnOffAudioRecording();
+        otherInformationPage.EnterOtherInformation(bookingDto.OtherInformation);
+
+        return otherInformationPage.GoToSummaryPage();
     }
 }
